controllers/transaction_controller.go
package controllers

import (
	"encoding/json"
	"net/http"

	"transactions/models"
	"transactions/services"
)

// CreateTransaction handles POST /transactions
func CreateTransaction(w http.ResponseWriter, r *http.Request) {
	var transaction models.Transaction
	if err := json.NewDecoder(r.Body).Decode(&transaction); err != nil {
		http.Error(w, "Invalid request payload", http.StatusBadRequest)
		return
	}

	createdTransaction, err := services.CreateTransaction(transaction)
	if err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(createdTransaction)
}
